agent/pyproject.toml
[project]
name = "proverbs-agent"
version = "0.1.0"
description = "ADK Proverbs Agent with shared state"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "python-dotenv",
    "pydantic",
    "google-adk==1.27.1",
    "google-genai",
    "ag-ui-adk",
    "python-docx",
    "pymupdf",
]
